Update README.md
This is my complete Internship work done in four weeks.

 WEEK ONE AND TWO:
•	Introduction to Google Collab 
•	GitHub installation
•	Introduction to Python
•	Python programming (Conditional Statements, Looping Statements, Functions, Lambda Functions)
•	Python Programming (Docstring, Error Handling, Lists)
•	Python Programming (Tuples, Sets, and Dictionaries)
•	Useful Libraries for Data Science in python
•	Getting start with NumPy
•	Getting start with Pandas
•	Getting start with seaborn
•	Data Processing
•	Data Visualization
•	Planetary Dataset Case Study
•	Covid Dataset Case Study
•	Solving Regression Problem

WEEK THREE AND FOUR:
the company assigned us some assignments which we have to solve in 15 days.
ASSIGNMENT 1:
Fish Dataset :
The dataset consists of 7 records (Species, Weight, Length1, Length2, Length3, Height, Width) with common different fish species which are available in the market. by analyzing this data, we can predict the estimated weight of fish. using this dataset we applied the concepts of data processing, data visualization, data analysis, and retrieved r2 score.